+++ modules/shadertools/src/modules/lighting/gouraud-material/gouraud-material.ts
@@ -4,6 +4,7 @@
 
 import {NumberArray3} from '@math.gl/types';
 import {ShaderModule} from '../../../lib/shader-module/shader-module';
+import {floatColors} from '../../color/float-colors';
 import {lighting} from '../lights/lighting';
 import {PHONG_VS, PHONG_FS} from '../phong-material/phong-shaders-glsl';
 import {PHONG_WGSL} from '../phong-material/phong-shaders-wgsl';
@@ -16,6 +17,8 @@ export type GouraudMaterialProps = {
   shininess?: number;
   specularColor?: [number, number, number];
 };
+
+const DEFAULT_SPECULAR_COLOR: NumberArray3 = [38.25, 38.25, 38.25];
 
 /** In Gouraud shading, color is calculated for each triangle vertex normal, and then color is interpolated colors across the triangle */
 export const gouraudMaterial: ShaderModule<GouraudMaterialProps> = {
@@ -30,7 +33,7 @@ export const gouraudMaterial: ShaderModule<GouraudMaterialProps> = {
   defines: {
     LIGHTING_VERTEX: true
   },
-  dependencies: [lighting],
+  dependencies: [lighting, floatColors],
   uniformTypes: {
     unlit: 'i32',
     ambient: 'f32',
@@ -43,14 +46,10 @@ export const gouraudMaterial: ShaderModule<GouraudMaterialProps> = {
     ambient: 0.35,
     diffuse: 0.6,
     shininess: 32,
-    specularColor: [0.15, 0.15, 0.15]
+    specularColor: DEFAULT_SPECULAR_COLOR
   },
 
   getUniforms(props: GouraudMaterialProps) {
-    const uniforms = {...props};
-    if (uniforms.specularColor) {
-      uniforms.specularColor = uniforms.specularColor.map(x => x / 255) as NumberArray3;
-    }
-    return {...gouraudMaterial.defaultUniforms, ...uniforms};
+    return {...gouraudMaterial.defaultUniforms, ...props};
   }
 };