+++ modules/shadertools/src/modules/lighting/phong-material/phong-material.ts
@@ -4,6 +4,7 @@
 
 import {NumberArray3} from '@math.gl/types';
 import {ShaderModule} from '../../../lib/shader-module/shader-module';
+import {floatColors} from '../../color/float-colors';
 import {lighting} from '../lights/lighting';
 import {PHONG_WGSL} from './phong-shaders-wgsl';
 import {PHONG_VS, PHONG_FS} from './phong-shaders-glsl';
@@ -17,12 +18,14 @@ export type PhongMaterialProps = {
   specularColor?: NumberArray3;
 };
 
+const DEFAULT_SPECULAR_COLOR: NumberArray3 = [38.25, 38.25, 38.25];
+
 /** In Phong shading, the normal vector is linearly interpolated across the surface of the polygon from the polygon's vertex normals. */
 export const phongMaterial: ShaderModule<PhongMaterialProps> = {
   name: 'phongMaterial',
   firstBindingSlot: 0,
   bindingLayout: [{name: 'phongMaterial', group: 3}],
-  dependencies: [lighting],
+  dependencies: [lighting, floatColors],
   // Note these are switched between phong and gouraud
   source: PHONG_WGSL,
   vs: PHONG_VS,
@@ -42,13 +45,9 @@ export const phongMaterial: ShaderModule<PhongMaterialProps> = {
     ambient: 0.35,
     diffuse: 0.6,
     shininess: 32,
-    specularColor: [0.15, 0.15, 0.15]
+    specularColor: DEFAULT_SPECULAR_COLOR
   },
   getUniforms(props?: PhongMaterialProps) {
-    const uniforms = {...props};
-    if (uniforms.specularColor) {
-      uniforms.specularColor = uniforms.specularColor.map(x => x / 255) as NumberArray3;
-    }
-    return {...phongMaterial.defaultUniforms, ...uniforms};
+    return {...phongMaterial.defaultUniforms, ...props};
   }
 };