+++ modules/shadertools/test/index.ts
@@ -41,6 +41,7 @@ import './modules/modules.spec';
 import './modules/utils/random.spec';
 
 // Lighting
+import './modules/color/float-colors.spec';
 import './modules/lighting/lights.spec';
 import './modules/lighting/dirlight.spec';
 import './modules/lighting/phong-material.spec';